fix: estrutura de los archivos
